refactor(memory): finish the facet cuts — graph wiring out of service.rs, compile helpers next to their consumer (#2036)
* refactor(memory): finish the facet cuts — graph wiring out of service.rs, compile helpers next to their consumer

The #2010 splits stopped at the budget line, not the seam line (#2021).

service.rs kept 21 S: GraphStore-bounded methods — the remember*/
autograph* family and the wire_*/store_extract* plumbing — while
service_graph.rs's own doc admitted the cut was partial. They now live in
service_graph_wiring.rs, a #[path] child of service like its siblings:
every GraphStore-bounded method is in one of the two service_graph*
modules (walks and destruction vs construction — either half alone would
blow the same file budget the split serves), and service.rs drops
1716 -> 1006 lines, baseline tightened accordingly. The worker pair
keeps its original non-wasm Send+Sync+'static impl block.

memory_bridge.rs held 10 free helpers consumed only by
memory_bridge_compile.rs; they now sit next to their one consumer and
the parent drops 940 -> 736 lines. Honest note against the issue's
estimate: the helpers were 194 lines (not ~330), and co-locating them
does not shrink the import seam to one line — they pull the CTX_* field
vocabulary with them, which stays shared with the parent's other
readers. source_id gains pub(super) for the sibling tests module, the
one adjustment the issue predicted.

Pure moves throughout: the diff is relocations plus the two visibility
markers and refreshed module docs; no semantics touched.

Closes #2021

* fix(memory): gate the worker-only import off wasm32

The wasm build check compiles velesdb-memory for wasm32 under
-D warnings; AutographWorkerHandle is consumed only by the spawned-worker
impl block that target compiles out, so its import must be gated with
the block. AutographJob stays unconditional — autograph_if's try_send
uses it on every target.
